+++ examples/cryptocurrency/src/lib.rs
@@ -41,6 +41,7 @@ pub mod proto;
 pub mod schema {
     use exonum::crypto::{Hash, PublicKey};
     use exonum_merkledb::{IndexAccess, MapIndex};
+    use exonum_proto::ProtobufConvert;
 
     use super::proto;
 
@@ -49,8 +50,8 @@ pub mod schema {
     //
     // [1]: https://exonum.com/doc/version/latest/architecture/serialization
     /// Wallet struct used to persist data within the service.
-    #[derive(Serialize, Deserialize, Clone, Debug, ProtobufConvert)]
-    #[exonum(pb = "proto::Wallet")]
+    #[derive(Serialize, Deserialize, Clone, Debug, ProtobufConvert, BinaryValue, ObjectHash)]
+    #[protobuf_convert(source = "proto::Wallet")]
     pub struct Wallet {
         /// Public key of the wallet owner.
         pub pub_key: PublicKey,
@@ -126,18 +127,19 @@ pub mod transactions {
     use exonum::crypto::PublicKey;
 
     use super::proto;
+    use exonum_proto::ProtobufConvert;
 
     /// Service configuration parameters.
-    #[derive(Serialize, Deserialize, Clone, Debug, ProtobufConvert)]
-    #[exonum(pb = "proto::Config")]
+    #[derive(Serialize, Deserialize, Clone, Debug, ProtobufConvert, BinaryValue, ObjectHash)]
+    #[protobuf_convert(source = "proto::Config")]
     pub struct Config;
 
     /// Transaction type for creating a new wallet.
     ///
     /// See [the `Transaction` trait implementation](#impl-Transaction) for details how
     /// `TxCreateWallet` transactions are processed.
-    #[derive(Serialize, Deserialize, Clone, Debug, ProtobufConvert)]
-    #[exonum(pb = "proto::TxCreateWallet")]
+    #[derive(Serialize, Deserialize, Clone, Debug, ProtobufConvert, BinaryValue, ObjectHash)]
+    #[protobuf_convert(source = "proto::TxCreateWallet")]
     pub struct TxCreateWallet {
         /// UTF-8 string with the owner's name.
         pub name: String,
@@ -147,8 +149,8 @@ pub mod transactions {
     ///
     /// See [the `Transaction` trait implementation](#impl-Transaction) for details how
     /// `TxTransfer` transactions are processed.
-    #[derive(Serialize, Deserialize, Clone, Debug, ProtobufConvert)]
-    #[exonum(pb = "proto::TxTransfer")]
+    #[protobuf_convert(source = "proto::TxTransfer")]
+    #[derive(Serialize, Deserialize, Clone, Debug, ProtobufConvert, BinaryValue, ObjectHash)]
     pub struct TxTransfer {
         /// Public key of the receiver.
         pub to: PublicKey,